requirements.txt
discord==2.2.2
Pillow==9.4.0
requests==2.28.2
yt-dlp==2023.3.4